requirements.txt
Flask==3.1.0
python-dotenv==1.1.0
python_bcrypt==0.3.2
Requests==2.32.3
